Revert "Scoped and Namespaced Clients (#20)"
This reverts commit b2a5008b9e80b8b9cd0184857b85860f2c316393.
