+++ src/K8sOperator.NET/Builder/EventWatcherBuilder.cs
@@ -1,5 +1,4 @@
-﻿using K8sOperator.NET.Metadata;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace K8sOperator.NET.Builder;
 
@@ -8,15 +7,6 @@ internal class EventWatcherBuilder(IServiceProvider serviceProvider, IController
     public IEventWatcher Build()
     {
         var watcherType = typeof(EventWatcher<>).MakeGenericType(controller.ResourceType);
-        IKubernetesClient client = ActivatorUtilities.CreateInstance<NamespacedKubernetesClient>(serviceProvider);
-
-        var clientScope = controller.ResourceType.GetCustomAttributes(false).OfType<IEntityScopeMetadata>().FirstOrDefault();
-        if (clientScope?.Scope == EntityScope.Cluster)
-        { 
-            client = ActivatorUtilities.CreateInstance<ClusterKubernetesClient>(serviceProvider);
-        }
-
-        
-        return (IEventWatcher)ActivatorUtilities.CreateInstance(serviceProvider, watcherType, client, controller, metadata);
+        return (IEventWatcher)ActivatorUtilities.CreateInstance(serviceProvider, watcherType, controller, metadata);
     }
 }
